Error messages on adding students to sections, add multiple students to a section at once

+++ public/javascripts/Controllers/teacherHomeCtrl.js
@@ -10,7 +10,9 @@ function($scope, auth, sections) {
   };
 
   $scope.deleteSection = function(section) {
-    sections.deleteSection(section);
+    if (confirm("Are you sure you want to delete " + section.title + "?")) {
+			sections.deleteSection(section);
+		}
   };
 
 }]);
